LLM_powered_Schema/LLmQuery/requirement.txt
Django
python-dotenv
langchain
langchain-openai
langchain-experimental
openai
sqlite3
git log -p
